backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy>=2.0.0
chromadb==0.4.18
google-generativeai==0.3.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
aiohttp==3.9.1
Pillow>=10.2.0
python-slugify==8.0.1
email_validator>=2.0.0
apscheduler>=3.10.0


